+++ internal/helpers/strings.go
@@ -17,6 +17,18 @@ func StringArraysEqual(a []string, b []string) bool {
 	return true
 }
 
+func StringArrayArraysEqual(a [][]string, b [][]string) bool {
+	if len(a) != len(b) {
+		return false
+	}
+	for i, x := range a {
+		if !StringArraysEqual(x, b[i]) {
+			return false
+		}
+	}
+	return true
+}
+
 func StringArrayToQuotedCommaSeparatedString(a []string) string {
 	sb := strings.Builder{}
 	for i, str := range a {